requirements.txt
matplotlib==3.8.4
nltk==3.8.1
numpy==1.26.4
Pillow==10.3.0
pycocotools==2.0.7
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
